Disable "puti" instruction
After testing, it seems the hardware decodes "puti" as another "exti",
since changing any "exti" instruction's opcode from 6 to 7 results in
the same operation being performed. To avoid confusion and incorrect
disassembly, let's disable decoding the "puti" instruction. That way, if
one is ever encountered in a real binary, it can be caught and the issue
investigated further.